refactor: extract repetition types to standalone Repetition file and update imports
