+++ __test__/components/RepeatEvent.test.tsx
@@ -2,7 +2,7 @@ import { RootState } from '@common/app/store'
 import RepeatEvent from '@common/components/Event/EventRepeat'
 import * as eventThunks from '@common/features/Calendars/CalendarSlice'
 import EventPopover from '@common/features/Events/EventModal'
-import { RepetitionObject } from '@common/types/EventsTypes'
+import { RepetitionObject } from '@common/types/Repetition'
 import { CalendarApi, DateSelectArg } from '@fullcalendar/core'
 import { act, fireEvent, screen, waitFor } from '@testing-library/react'
 import { renderWithProviders } from '../utils/Renderwithproviders'
