+++ common/src/components/Event/EventRepeat.tsx
@@ -1,4 +1,4 @@
-import { RepetitionObject } from '@common/types/EventsTypes'
+import { RepetitionObject } from '@common/types/Repetition'
 import {
   Box,
   FormControl,
