+++ common/src/components/Event/components/DateTimeSummary.tsx
@@ -8,7 +8,7 @@ import 'dayjs/locale/vi'
 import React from 'react'
 import { useI18n } from 'twake-i18n'
 import { getTimezoneOffset } from '@common/utils/timezone'
-import { RepetitionObject } from '@common/types/EventsTypes'
+import { RepetitionObject } from '@common/types/Repetition'
 import { LONG_DATE_FORMAT } from '@common/components/Event/utils/dateTimeFormatters'
 import { SectionPreviewRow } from './SectionPreviewRow'
 import { makeRecurrenceString } from '@common/components/EventPreview/utils/makeRecurrenceString'
