+++ common/src/components/Event/utils/initialValues/recurrenceResolvers.ts
@@ -1,5 +1,6 @@
 import { Calendar } from '@common/types/CalendarTypes'
-import { CalendarEvent, RepetitionObject } from '@common/types/EventsTypes'
+import { CalendarEvent } from '@common/types/EventsTypes'
+import { RepetitionObject } from '@common/types/Repetition'
 import { extractEventBaseUuid } from '@common/utils/extractEventBaseUuid'
 
 function getBaseRepetition(
