+++ common/src/components/EventPreview/EventDetailsRows.tsx
@@ -2,7 +2,7 @@ import { VideoLink } from '@common/components/Event/components/VideoLink'
 import { InfoRow } from '@common/components/Event/InfoRow'
 import { userAttendee } from '@common/features/User/models/attendee'
 import { Attachment } from '@common/types/Attachment'
-import { RepetitionObject } from '@common/types/EventsTypes'
+import { RepetitionObject } from '@common/types/Repetition'
 import { Valarms } from '@common/types/Valarms'
 import { Box, Typography, useTheme } from '@linagora/twake-mui'
 import ErrorOutlinedIcon from '@mui/icons-material/ErrorOutlined'
