+++ common/src/features/Events/hooks/submitCreateHelpers/createAction.ts
@@ -6,7 +6,8 @@ import { putEvent } from '@common/features/Calendars/CalendarSlice'
 import { buildDelegatedEventURL } from '@common/features/Events/utils/buildDelegatedEventURL'
 import { userAttendee } from '@common/features/User/models/attendee'
 import { Calendar } from '@common/types/CalendarTypes'
-import { CalendarEvent, RepetitionObject } from '@common/types/EventsTypes'
+import { CalendarEvent } from '@common/types/EventsTypes'
+import { RepetitionObject } from '@common/types/Repetition'
 import { Valarms } from '@common/types/Valarms'
 import { assertThunkSuccess } from '@common/utils/assertThunkSuccess'
 import {
