+++ common/src/types/EventsTypes.ts
@@ -1,14 +1,11 @@
 import { Resource } from '@common/components/Attendees/ResourceSearch'
 import { Calendar } from '@common/types/CalendarTypes'
-import {
-  RepetitionRule,
-  VObjectProperty
-} from '@common/features/Calendars/types/CalendarData'
+import { VObjectProperty } from '@common/features/Calendars/types/CalendarData'
 import { userAttendee } from '@common/features/User/models/attendee'
 import { userOrganiser } from '@common/features/User/userDataTypes'
 import { Attachment } from './Attachment'
 import { Valarms } from './Valarms'
-import { formatUntilForRRule } from '@common/features/Events/utils/formatDateToICal'
+import { RepetitionObject } from './Repetition'
 
 export interface CalendarEvent {
   URL: string
@@ -40,61 +37,7 @@ export interface CalendarEvent {
   attach?: Attachment[]
 }
 
-export interface RepetitionData {
-  freq: string
-  interval?: number | null
-  byday?: string[] | null
-  occurrences?: number | null
-  endDate?: string | null
-  wkst?: string | null
-}
 
-export type RepetitionOptions = Partial<RepetitionData>
-
-export class RepetitionObject implements RepetitionData {
-  freq: string
-  interval?: number | null
-  byday?: string[] | null
-  occurrences?: number | null
-  endDate?: string | null
-  wkst?: string | null
-
-  constructor({
-    freq,
-    interval,
-    byday,
-    occurrences,
-    endDate,
-    wkst
-  }: RepetitionOptions = {}) {
-    this.freq = freq ?? ''
-    this.interval = interval
-    this.byday = byday
-    this.occurrences = occurrences
-    this.endDate = endDate
-    this.wkst = wkst
-  }
-
-  asJcal(allday: boolean, tzid: string): VObjectProperty {
-    const repetitionRule: RepetitionRule = { freq: this.freq }
-    if (this.interval != null) {
-      repetitionRule.interval = this.interval
-    }
-    if (this.occurrences != null) {
-      repetitionRule.count = this.occurrences
-    }
-    if (this.endDate) {
-      repetitionRule.until = formatUntilForRRule(this.endDate, allday, tzid)
-    }
-    if (this.byday !== null && this.byday !== undefined) {
-      repetitionRule.byday = this.byday
-    }
-    if (this.wkst) {
-      repetitionRule.wkst = this.wkst
-    }
-    return ['rrule', {}, 'recur', repetitionRule]
-  }
-}
 
 export interface ContextualizedEvent {
   event: CalendarEvent
